+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/assembly/KafkaBridgeAssemblyOperatorTest.java
@@ -157,7 +157,7 @@ public class KafkaBridgeAssemblyOperatorTest {
 
             // No metrics config  => no CMs created
             Set<String> metricsNames = new HashSet<>();
-            if (bridge.isPrometheusMetricsEnabled()) {
+            if (bridge.isMetricsEnabled()) {
                 metricsNames.add(KafkaBridgeResources.metricsAndLogConfigMapName(clusterCmName));
             }