+++ cluster-operator/src/test/java/io/strimzi/operator/cluster/operator/assembly/PartialRollingUpdateTest.java
@@ -96,7 +96,7 @@ public class PartialRollingUpdateTest {
                             .withStorageClass("foo")
                             .withDeleteClaim(true)
                         .endPersistentClaimStorage()
-                        .withPrometheusMetrics(emptyMap())
+                        .withMetrics(emptyMap())
                     .endKafka()
                     .withNewZookeeper()
                         .withReplicas(3)